add async handler and global handling error and authentication with tokens and hashing password

+++ src/modules/note/note.controller.js
@@ -1,87 +1,69 @@
 import { Note } from "../../../DB/models/note.model.js"
 import { User } from "../../../DB/models/user.model.js"
+import { asyncHandler } from "../../utils/asyncHandler.js"
 //create note
-export const createNote =async(req,res)=>{
-    try {
-        const {content,user}=req.body
-        //check user 
-        const isUser = await User.findById(user)
-        if(!isUser) return res.json({success:false ,message:"you can't add note!!!"})
+export const createNote = asyncHandler(async(req,res,next)=>{
+   
+    const user = req.user._id
+        const {content}=req.body
+
 
         //create note
         await Note.create({content , user})
         //send res
         res.json({success:true , message:"note added successfully"})
 
-    } catch (error) {
-        return res.json({success:false, message :"here in catch", error})
-    }
-}
+})
 
 // update note 
-export const updateNote = async(req,res)=>{
-    try {
+export const updateNote = asyncHandler(async(req,res,next)=>{
+   
         const {id}=req.params
-        const {user,isCompleted}=req.body
-        // check user 
-        const isUser = await User.findById(user)
-        if(!isUser) return res.json({success:false ,message:"you can't update note!!!"})
+        const user = req.user._id
+        const {isCompleted}=req.body
+        
         // check note 
     const note = await Note.findById(id)
     console.log(note);
-    if(!note)res.json({sucess:false ,message : "note not found "})
+    if(!note) return next( new Error("note not found"))
     //check owner
     if(note.user.toString()!==user.toString()) 
-        return res.json({sucess:false ,message : "your not the owner "})
+        return next( new Error("you are not the owner"))
         //update note
         note.isCompleted =isCompleted
         await note.save()
         res.json({success:true , message:"note updated successfully"})
-    } catch (error) {
-        return res.json({success:false, message :"here in catch", error})
-    }
-}
 
-export const deleteNote= async(req,res)=>{
-    try {
+})
+
+export const deleteNote=  asyncHandler(async(req,res,next)=>{
+
         const {id}=req.params
-        const {user}=req.body
-        // check user 
-        const isUser = await User.findById(user)
-        if(!isUser) return res.json({success:false ,message:"you can't delete note!!!"})
+        const user = req.user._id
+
         // check note 
     const note = await Note.findById(id)
     console.log(note);
-    if(!note)res.json({sucess:false ,message : "note not found "})
+    if(!note) return next( new Error("note not found"))
     //check owner
     if(note.user.toString()!==user.toString()) 
-    return res.json({sucess:false ,message : "your not the owner "})
+    return next( new Error("you are not the owner"))
     //delete note
     await note.deleteOne()
     res.json({success:true , message:"note deleted successfully"})
 
-} catch (error) {
-    return res.json({success:false, message :"here in catch", error})
-}
-}
-export const allNotes =async(req,res)=>{
-    try {
+
+})
+export const allNotes = asyncHandler(async(req,res)=>{
+   
          const notes = await Note.find().populate("user")
          return res.json({success:true , results :{notes}})
-} 
-catch (error) {
-    return res.json({success:false, message :"here in catch", error})
-}
-}
-export const userNotes = async(req,res)=>{
-try {
+})
+export const userNotes = asyncHandler(async(req,res)=>{
 const {id} = req.params
         // check user 
         const isUser = await User.findById(id)
-        if(!isUser) return res.json({success:false ,message:"you can't delete note!!!"})
+        if(!isUser) return next( new Error("you are not a user "))
         const notes = await Note.find({user:id})
         return res.json({success:true , results:{notes}})
-} catch (error) {
-    return res.json({success:false, message :"here in catch", error})
-}
-}
+})
